+++ R/criterion.R
@@ -10,22 +10,64 @@
 #'   if the directory specified by this parameter is the project root,
 #'   and \code{FALSE} otherwise
 #' @param desc A textual description of the test criterion
+#' @param subdir Subdirectories to start the search in, if found
+#'
+#' @return
+#' An S3 object of class \code{root_criterion} wit the following members:
 #'
 #' @include rrmake.R
 #' @export
-root_criterion <- function(testfun, desc) {
+#'
+#' @examples
+#' root_criterion(function(path) file.exists(file.path(path, "somefile")), "Has somefile")
+#' has_file("DESCRIPTION")
+#' is_r_package
+#' is_r_package$find_file
+#' \dontrun{
+#' is_r_package$make_fix_file(".")
+#' }
+root_criterion <- function(testfun, desc, subdir = NULL) {
   if (!isTRUE(all.equal(names(formals(testfun)), "path"))) {
     stop("testfun must be a function with one argument 'path'")
   }
+
+  full_desc <- paste0(
+    desc,
+    if (!is.null(subdir)) paste0(
+      " (also look in subdirectories: ",
+      paste0("'", subdir, "'", collapse = ", "),
+      ")"
+    )
+  )
+
   criterion <- structure(
     list(
+      #' @return
+      #' \describe{
+      #'   \item{\code{testfun}}{The \code{testfun} argument}
       testfun = testfun,
-      desc = desc
+      #'   \item{\code{desc}}{The \code{desc} argument}
+      desc = full_desc,
+      #'   \item{\code{subdir}}{The \code{subdir} argument}
+      subdir = subdir
     ),
     class = "root_criterion"
   )
 
+  #'   \item{\code{find_file}}{A function with \code{...} argument that returns
+  #'     for a path relative to the root specified by this criterion.
+  #'     The optional \code{path} argument specifies the starting directory,
+  #'     which defaults to \code{"."}.
+  #'   }
   criterion$find_file <- make_find_root_file(criterion)
+  #'   \item{\code{make_fix_file}}{A function with a \code{path} argument that
+  #'      returns a function that finds paths relative to the root.  For a
+  #'      criterion \code{cr}, the result of \code{cr$make_fix_file(".")(...)}
+  #'      is identical to \code{cr$find_file(...)}. The function created by
+  #'      \code{make_fix_file} can be saved to a variable to be more independent
+  #'      of the current working directory.
+  #'   }
+  #' }
   criterion$make_fix_file <-
     function(path = getwd()) make_fix_root_file(criterion, path)
 
@@ -71,4 +113,14 @@ format.root_criterion <- function(x, ...) {
 #' @export
 print.root_criterion <- function(x, ...) {
   cat(paste0(format(x), "\n"))
+}
+
+#' @export
+`|.root_criterion` <- function(x, y) {
+  stopifnot(is.root_criterion(y))
+
+  root_criterion(
+    function(path) x$testfun(path) || y$testfun(path),
+    paste0(x$desc, ", or ", y$desc)
+  )
 }