+++ R/has-file.R
@@ -66,11 +66,39 @@ has_file_pattern <- function(pattern, contents = NULL, n = -1L) {
   root_criterion(testfun, desc)
 }
 
+#' @details
+#' The \code{has_dirname} function constructs a criterion that checks if the
+#' \code{\link[base]{dirname}} has a specific name.
+#'
+#' @rdname root_criterion
+#' @param dirname A directory name, without subdirectories
+#' @export
+has_dirname <- function(dirname, subdir = NULL) {
+  force(dirname)
+
+  testfun <- eval(bquote(function(path) {
+    dir.exists(file.path(dirname(path), .(dirname)))
+  }))
+
+  desc <- paste0("Directory name is '", dirname, "'")
+
+  root_criterion(testfun, desc, subdir = subdir)
+}
+
 #' @export
 is_rstudio_project <- has_file_pattern("[.]Rproj$", contents = "^Version: ", n = 1L)
 
 #' @export
 is_r_package <- has_file("DESCRIPTION", contents = "^Package: ")
+
+#' @export
+is_remake_project <- has_file("remake.yml")
+
+#' @export
+from_wd <- root_criterion(function(path) TRUE, "From current working directory")
+
+#' @export
+is_testthat <- has_dirname("testthat", c("tests/testthat", "testthat"))
 
 #' Prespecified criteria
 #'
@@ -80,7 +108,10 @@ is_r_package <- has_file("DESCRIPTION", contents = "^Package: ")
 criteria <- structure(
   list(
     is_rstudio_project = is_rstudio_project,
-    is_r_package = is_r_package
+    is_r_package = is_r_package,
+    is_remake_project = is_remake_project,
+    is_testthat = is_testthat,
+    from_wd = from_wd
   ),
   class = "root_criteria")
 
@@ -104,6 +135,28 @@ str.root_criteria <- function(object, ...) {
 #' @export
 "is_r_package"
 
+#' @details
+#' \code{is_remake_project} looks for a \code{remake.yml} file.
+#'
+#' @rdname criteria
+#' @export
+"is_remake_project"
+
+#' @details
+#' \code{is_testthat} looks for the \code{testthat} directory, works when
+#'   developing, testing, and checking a package.
+#'
+#' @rdname criteria
+#' @export
+"is_testthat"
+
+#' @details
+#' \code{from_wd} uses the current working directory.
+#'
+#' @rdname criteria
+#' @export
+"from_wd"
+
 
 list_files <- function(path, filename) {
   files <- dir(path = path, pattern = filename, all.files = TRUE, full.names = TRUE)
@@ -113,7 +166,7 @@ list_files <- function(path, filename) {
 }
 
 is_dir <- function(x) {
-  file.info(x, extra_cols = FALSE)$isdir
+  dir.exists(x)
 }
 
 match_contents <- function(f, contents, n) {